requirements.txt
alembic==1.4.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
dataset==1.3.2
decorator==4.4.2
idna==2.10
Mako==1.1.3
MarkupSafe==1.1.1
praw==7.1.0
prawcore==1.4.0
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
python-telegram-bot==12.8
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
tornado==6.0.4
update-checker==0.17
urllib3==1.25.9
websocket-client==0.57.0
youtube-dl==2020.6.16.1
